# swagger.yaml
{"swagger":"2.0","info":{"description":"Crud","version":"2.0","title":"Fast-Track API Documentation","contact":{"name":"Fabio Alves","email":"fabio.calves@telefonica.com"}},"host":"localhost:9999","basePath":"/","tags":[{"name":"order-controller","description":"Order Controller"}],"paths":{"/orders":{"get":{"tags":["order-controller"],"summary":"list","operationId":"listUsingGET","produces":["*/*"],"parameters":[{"name":"name","in":"query","description":"name","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/OrderDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["order-controller"],"summary":"create","operationId":"createUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"orderIn","description":"orderIn","required":true,"schema":{"$ref":"#/definitions/OrderIn"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/OrderDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/orders/search":{"get":{"tags":["order-controller"],"summary":"search","operationId":"searchUsingGET","produces":["*/*"],"parameters":[{"name":"max_total","in":"query","required":false,"type":"string"},{"name":"maxTotaldb","in":"query","required":false,"type":"number","format":"double"},{"name":"min_total","in":"query","required":false,"type":"string"},{"name":"minTotaldb","in":"query","required":false,"type":"number","format":"double"},{"name":"q","in":"query","required":false,"type":"string"},{"name":"valid","in":"query","required":false,"type":"boolean"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/OrderDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/orders/{id}":{"get":{"tags":["order-controller"],"summary":"findById","operationId":"findByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/OrderDto"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["order-controller"],"summary":"update","operationId":"updateUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"in":"body","name":"updateOrderIn","description":"updateOrderIn","required":true,"schema":{"$ref":"#/definitions/UpdateOrderIn"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/OrderDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["order-controller"],"summary":"delete","operationId":"deleteUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}}},"definitions":{"OrderDto":{"type":"object","properties":{"description":{"type":"string"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"status":{"type":"string","enum":["NOT_PROCESSED","PROCESSED"]},"total":{"type":"number","format":"double"}},"title":"OrderDto"},"OrderIn":{"type":"object","properties":{"description":{"type":"string"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"status":{"type":"string","enum":["NOT_PROCESSED","PROCESSED"]},"total":{"type":"number","format":"double"}},"title":"OrderIn"},"UpdateOrderIn":{"type":"object","properties":{"description":{"type":"string"},"name":{"type":"string"},"status":{"type":"string","enum":["NOT_PROCESSED","PROCESSED"]},"total":{"type":"number","format":"double"}},"title":"UpdateOrderIn"}}}